modify config file format.
