+++ dbyml/data/full.yml
@@ -1,27 +1,30 @@
 ---
-name: dbyml-sample
-tag: latest
-path: .
-dockerfile: Dockerfile
-target: ""
-build_args:
-  key1: value1
-  key2: value2
-label:
-  label1: label1
-  label2: label2
-stdout: true
-no_cache: false
-remove_intermediate_container: true
-force_rm: true
-remove_dangling: true
-push:
-  enabled: true
-  username: ${REGISTRY_USER}
-  password: ${REGISTRY_PASSWORD}
-  registry:
-    host: "myregistry.com"
-    port: "5000"
+image:
+  name: dbyml-sample
+  tag: latest
+  path: .
+  dockerfile: Dockerfile
+  build_args:
+    key1: value1
+    key2: value2
+  label:
+    label1: label1
+    label2: label2
+
+build:
+  target: ""
+  stdout: true
+  no_cache: false
+  remove_intermediate_container: true
+  force_rm: true
+  remove_dangling: true
+  verbose: true
+
+registry:
+  enabled: false
+  username: ""
+  password: ""
+  host: "myregistry.com"
+  port: 5000
   namespace: public
   remove_local: true
-verbose: true
