--- dbyml/data/normal.yml
@@ -1,16 +1,0 @@
----
-name: dbyml-sample
-tag: latest
-path: .
-dockerfile: Dockerfile
-build_args:
-  key1: value1
-  key2: value2
-label:
-  label1: label1
-  label2: label2
-stdout: true
-no_cache: false
-remove_intermediate_container: true
-force_rm: true
-remove_dangling: true
