+++ sample/sample.yml
@@ -4,74 +4,81 @@
 # All options are listed.
 #
 
-# name: Image name
-name: dbyml-sample
+# The image filed manages docker image attributes.
+image:
+  # name: Image name. This field is required.
+  name: dbyml-sample
 
-# tag: Image tag
-tag: latest
+  # tag: Image tag.
+  tag: latest
 
-# path: Path to a directory containing Dockerfile.
-path: sample/dockerfiles
+  # path: Path to a directory containing Dockerfile.
+  path: sample/dockerfiles
 
-# dockerfile: Filename of Dockerfile.
-dockerfile: Dockerfile
+  # dockerfile: Filename of Dockerfile.
+  dockerfile: Dockerfile
 
-# target: Name of the build-stage to build in a multi-stage Dockerfile.
-target: ""
+  # build_args: Arguments corresponding to build-args when use docker build.
+  # Set key: value
+  build_args:
+    key1: value1
+    key2: value2
 
-# build_args: Arguments corresponding to build-args when use docker build.
-# Set key: value
-build_args:
-  key1: value1
-  key2: value2
+  # label: Arguments corresponding to label when use docker build.
+  # Use quotation when the key name includes "." such as com.test.name
+  label:
+    aaa: this is test
+    build_info: test
+    "com.test.name": test name
 
-# label: Arguments corresponding to label when use docker build.
-# Use quotation when the key name includes "." such as com.test.name
-label:
-  aaa: this is test
-  build_info: test
-  "com.test.name": test name
 
-# stdout: If set true, show messages while building the image to stdout.
-# Set false to no output during build.
-# default: true
-stdout: true
+# The build field manages messages, intermediates and options on build.
+build:
+  # target: Name of the build-stage to build in a multi-stage Dockerfile.
+  target: ""
 
-# no_cache: Set true not to use build cache when build the image.
-# default: false
-no_cache: false
+  # stdout: If set true, show status messages when build and push the image to stdout.
+  # default: true
+  stdout: true
 
-# remove_intermediate_container: Set true to remove intermediate containers created during build.
-# default: true
-remove_intermediate_container: true
+  # no_cache: Set true not to use build cache when build the image.
+  # default: false
+  no_cache: false
 
-# force_rm: Set true to always remove intermediate containers, even after unsuccessful builds
-# default: true
-force_rm: true
+  # remove_intermediate_container: Set true to remove intermediate containers created during build.
+  # default: true
+  remove_intermediate_container: true
 
-# remove_dangling: Set true to delete the image with the same name:tag as the image to be built from local before building.
-# default: false
-remove_dangling: true
+  # force_rm: Set true to always remove intermediate containers, even after unsuccessful builds
+  # default: true
+  force_rm: true
 
-# push fields are required when pushing the image to be built to a registry.
-push:
+  # remove_dangling: Set true to delete the image with the same name:tag as the image to be built from local before building.
+  # default: false
+  remove_dangling: true
+
+  # verbose: Set true to show build settings.
+  # default: true
+  verbose: true
+
+
+# The registry field manages the information about registry to which build image push.
+registry:
   # enabled: Enable push to a registry. Set false not to push the image
   # to the registry even if the push fields are set.
   enabled: true
+
+  # host: Registry name or ip address.
+  host: "myregistry.com"
+
+  # port: Registry port.
+  port: "5000"
 
   # username: Registry auth username
   username: ${REGISTRY_USER}
 
   # password: Registry auth password
   password: ${REGISTRY_PASSWORD}
-
-  # registry: A registry to which push the image.
-  registry:
-    # host: Registry name or ip address.
-    host: "myregistry.com"
-
-    # port: Registry port.
-    port: "5000"
 
   # namespace:
   # When set the namespace, the image to be pushed to the registry will be {host}:{port}/{namespace}/{name}:{tag},
@@ -80,7 +87,3 @@ push:
 
   # remove_local: Set true to delete the image from local after successfully push to the registry.
   remove_local: false
-
-# verbose: Set true to show build settings.
-# default: false
-verbose: true
