--- tests/config/auth.yml
@@ -1,24 +1,0 @@
----
-name: dbyml-sample
-tag: latest
-path: tests/config/dockerfiles
-dockerfile: Dockerfile
-build_args:
-  key1: test1
-  key2: test2
-label:
-  aaa: "this is aaa"
-  build_dir: this is it
-  "com.test.aaa": aaa
-remove_dangling: true
-
-push:
-  enabled: true
-  username: docker
-  password: docker
-  registry:
-    host: "0.0.0.0"
-    port: 5551
-  remove_local: false
-
-verbose: true
