+++ tests/config/dbyml.yml.old
@@ -1,0 +1,23 @@
+---
+name: dbyml-sample
+tag: latest
+path: tests/config/dockerfiles
+dockerfile: Dockerfile
+target: ''
+label:
+  build_dir: ${PWD}
+stdout: true
+no_cache: false
+remove_intermediate_container: true
+force_rm: true
+remove_dangling: false
+push:
+  enabled: true
+  # username: ${REGISTRY_USER}
+  # password: ${REGISTRY_PASSWORD}
+  registry:
+    host: "0.0.0.0"
+    port: 5550
+  namespace: public
+  remove_local: true
+verbose: true
