+++ tests/config/env.yml
@@ -1,10 +1,12 @@
 ---
-name: dbyml-sample
-tag: latest
-path: tests/config/dockerfiles
-dockerfile: Dockerfile
-label:
-  env1: ${ENV1}
-  env_default: ${ENV_DEFAULT:-default_value}
-  multi_env: "${ENV1}/test/${ENV2}.${ENV3}"
-remove_dangling: true
+image:
+  name: dbyml-sample
+  tag: latest
+  path: tests/config/dockerfiles
+  dockerfile: Dockerfile
+  label:
+    env1: ${ENV1}
+    env_default: ${ENV_DEFAULT:-default_value}
+    multi_env: "${ENV1}/test/${ENV2}.${ENV3}"
+build:
+  remove_dangling: true
